Scope back to energy-only: revert hand-written power to float
The original branch globally flipped both code generators and hand-edited
code to make all numeric values double. Per the ticket, only energy (Wh)
values need double precision. Generated types stay double (JSON-serialized,
no float/double distinction on the wire; selective revert would need a
fragile field-name heuristic), and the OCPP library stays double (its
hand-written smart-charging code is type-coupled to the generated types).

This reverts the hand-written, non-generated power/current/voltage code back
to float: everest_api_types API structs (keeping powermeter energy_Wh as
double), iso15118/EvseV2G decoded physical values, and EnergyManager/
EnergyNode trading math. Locals that only mirror generated double fields
(EvseManager voltage, EnergyNode current derived from double power/voltage)
stay double to avoid fighting the generated types. Fixture 9 is regenerated
to the resulting float-precision power output.

+++ lib/everest/everest_api_types/include/everest_api_types/energy/API.hpp
@@ -12,7 +12,7 @@
 namespace everest::lib::API::V1_0::types::energy {
 
 struct NumberWithSource {
-    double value;
+    float value;
     std::string source;
 };
 
@@ -22,21 +22,21 @@ struct IntegerWithSource {
 };
 
 struct FrequencyWattPoint {
-    double frequency_Hz;
-    double total_power_W;
+    float frequency_Hz;
+    float total_power_W;
 };
 
 struct SetpointType {
     int32_t priority;
     std::string source;
-    std::optional<double> ac_current_A;
-    std::optional<double> total_power_W;
+    std::optional<float> ac_current_A;
+    std::optional<float> total_power_W;
     std::optional<std::vector<types::energy::FrequencyWattPoint>> frequency_table;
 };
 
 struct PricePerkWh {
     std::string timestamp;
-    double value;
+    float value;
     std::string currency;
 };
 
@@ -60,7 +60,7 @@ struct ScheduleReqEntry {
     std::string timestamp;
     LimitsReq limits_to_root;
     LimitsReq limits_to_leaves;
-    std::optional<double> conversion_efficiency;
+    std::optional<float> conversion_efficiency;
     std::optional<PricePerkWh> price_per_kwh;
 };
 
@@ -89,10 +89,10 @@ struct EnforcedLimits {
 };
 
 struct CapabilityLimits {
-    double max_current_A;
+    float max_current_A;
     int32_t max_phase_count;
-    double nominal_voltage_V;
-    double total_power_W;
+    float nominal_voltage_V;
+    float total_power_W;
 };
 
 } // namespace everest::lib::API::V1_0::types::energy
